backend/requirements.txt
fastapi
uvicorn
pandas
plotly
python-multipart
google-generativeai
python-dotenv
pillow
